ci: format code
