+++ src/config/src/utils/mod.rs
@@ -18,7 +18,6 @@ pub mod async_file;
 pub mod async_walkdir;
 pub mod base64;
 pub mod download_utils;
-pub mod encrypt;
 pub mod encryption;
 pub mod enrichment_local_cache;
 pub mod file;
